refactor: Move _convert_osmdict_to_gdf to src
